requirements.txt
altgraph==0.17.4
asgiref==3.5.0
bidict==0.22.0
brotlipy==0.7.0
certifi==2020.12.5
conda==4.9.2
Django==4.0.7
django-sass-processor==1.4.1
dnspython==2.2.1
gevent==24.10.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
importlib_metadata==8.5.0
netifaces==0.11.0
packaging==24.1
psutil==6.1.0
pycosat==0.6.3
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
python-engineio==4.3.1
python-socketio==5.5.2
simple-websocket==1.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
wsproto==1.2.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0
